Use native Int in UMFPACK control tests for 32-bit.
SparseArrays only defines get_umfpack_control for the platform Int
index type; hard-coded Int64 fails on i686.

+++ test/Core/umfpack_control.jl
@@ -54,13 +54,13 @@ using LinearSolve, SparseArrays, LinearAlgebra, Test, Random
         # Built directly rather than through a solve so the probe value never has
         # to be a legal setting for that particular knob.
         @testset "every documented setting maps to its own entry" begin
-            reference = get_umfpack_control(Float64, Int64)
+            reference = get_umfpack_control(Float64, Int)
             for setting in LinearSolve.UMFPACK_CONTROL_KEYS
                 idx = LinearSolve._UMFPACK_CONTROL_INDEX[setting]
                 probe = reference[idx] + 1
                 control = LinearSolve._umfpack_control(
                     UMFPACKFactorization(control = NamedTuple{(setting,)}((probe,))),
-                    Float64, Int64
+                    Float64, Int
                 )
                 @test control[idx] == probe
                 moved = findall(i -> control[i] != reference[i], eachindex(reference))
@@ -99,7 +99,7 @@ using LinearSolve, SparseArrays, LinearAlgebra, Test, Random
         end
 
         @testset "unnamed entries keep SparseArrays' defaults" begin
-            reference = get_umfpack_control(Float64, Int64)
+            reference = get_umfpack_control(Float64, Int)
             cache = init(
                 LinearProblem(copy(A), copy(b)),
                 UMFPACKFactorization(control = (; irstep = 2))
